Don't limit menu values to 8bits, fixes truncated bpm set

+++ app/inc/system.h
@@ -62,7 +62,7 @@
 
 // system menu callbacks
 void system_lock_comm_serial(bool busy);
-void system_update_menu_value(uint8_t item_ID, uint8_t value);
+void system_update_menu_value(uint8_t item_ID, uint16_t value);
 uint8_t system_get_current_profile(void);
 void system_true_bypass_cb(void *arg, int event);
 void system_pedalboard_cb(void *arg, int event);
